Fixed swiftLint issue with jobController.
